Reset non-closed circuitbreaker upon success
Even if a circuitbreaker is currently open, a success indicates that it
should be reset back to closed. This is what appears to have been
intended, it just looks like the original author didn't consider rare
race conditions in which a success could be achieved while the breaker
was open.