eliminate deadlocks from Subscribe and AddListener
The existing code provides two mechanisms for clients to receive events from
a circuit breaker. Both contained logic which could potentially lead to
deadlocks. The logic assumed that clients prefer to receive more recent
events in the case where their channel is full and thus attempts to pull
events off the front of the client channel in the case where sends would block.
This however can lead to deadlocking when another goroutine is trying to send
an event on the channel (which can totally happen). This commit fixes the
possible deadlocks, adds testing, and clarifies in the comments that the event
mechanisms are best effort. This course of action was taken as opposed to
removing the functionality so as to not break existing external clients who may
be relying on it.